Testing Direct Component Array Access

+++ src/StarlightEngine/Utils/TypeWrapper.h
@@ -2,8 +2,8 @@
 
 namespace Starlight
 {
-	using ObjectID = size_t;
-	using TypeID = size_t;	
+	using ObjectID = uint32_t;
+	using TypeID = uint32_t;
 	
 	using ComponentID = ObjectID;
 	using EntityID = ObjectID;
